Move VibitsAppUiState and SyncEffects to app/view
View state belongs with view components.
