+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/app/view/SyncEffects.kt
@@ -1,4 +1,4 @@
-package space.be1ski.vibits.shared.app
+package space.be1ski.vibits.shared.app.view
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
