+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/app/view/TimeRangeNavigation.kt
@@ -16,8 +16,6 @@ import kotlinx.datetime.LocalDate
 import kotlinx.datetime.Month
 import kotlinx.datetime.minus
 import kotlinx.datetime.plus
-import space.be1ski.vibits.shared.app.MemosScreen
-import space.be1ski.vibits.shared.app.VibitsAppUiState
 import space.be1ski.vibits.shared.core.platform.isDesktop
 import space.be1ski.vibits.shared.core.ui.ActivityMode
 import space.be1ski.vibits.shared.core.ui.ActivityRange
