+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/app/view/VibitsApp.kt
@@ -27,9 +27,6 @@ import org.jetbrains.compose.resources.stringResource
 import space.be1ski.vibits.shared.Res
 import space.be1ski.vibits.shared.action_create_memo
 import space.be1ski.vibits.shared.action_track_today
-import space.be1ski.vibits.shared.app.MemosScreen
-import space.be1ski.vibits.shared.app.SyncAutoLoad
-import space.be1ski.vibits.shared.app.VibitsAppUiState
 import space.be1ski.vibits.shared.app.di.VibitsAppDependencies
 import space.be1ski.vibits.shared.core.platform.currentLocalDate
 import space.be1ski.vibits.shared.core.ui.ActivityRange
