+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/app/view/VibitsAppComponents.kt
@@ -22,8 +22,6 @@ import androidx.compose.ui.Modifier
 import org.jetbrains.compose.resources.stringResource
 import space.be1ski.vibits.shared.Res
 import space.be1ski.vibits.shared.action_refresh
-import space.be1ski.vibits.shared.app.MemosScreen
-import space.be1ski.vibits.shared.app.VibitsAppUiState
 import space.be1ski.vibits.shared.app_name
 import space.be1ski.vibits.shared.core.platform.currentLocalDate
 import space.be1ski.vibits.shared.core.platform.isDesktop
